minor changes to End experiment and fixed adapter error
