+++ .gitignore
@@ -1,5 +1,4 @@
 *.iml
-.gradle
 /local.properties
 .idea
 .DS_Store
