+++ app/src/main/AndroidManifest.xml
@@ -35,7 +35,7 @@
                 android:value=".UI.main_menu.forum.ForumQuestionsActivity" />
         </activity>
           
-        <activity android:name=".UI.main_menu.my_experiments.MyExperimentActivity">
+        <activity android:name=".UI.main_menu.my_experiment.MyExperimentActivity">
 
         
         </activity>
@@ -64,7 +64,7 @@
 
         </activity>
   
-        <activity android:name=".UI.main_menu.my_experiments.PublishExpActivity">
+        <activity android:name=".UI.main_menu.my_experiment.PublishExpActivity">
 
         </activity>
